refactor: move multica hub metadata to surface package
Move Multica hub metadata, assignment fingerprints, and projection ledger state into the surface/multica adapter package while keeping driver facade helpers for existing issue types.

Validation: go test ./harness/internal/surface/multica ./harness/internal/driver ./harness/cmd/mnemon-multica-runtime ./harness/cmd/mnemon-acceptance ./harness/cmd/mnemon-harness
